Adds export integration test (#10)
* Adds export integration test

Adds a new integration test to verify the export functionality.

The test exports the genesis file from a running full node, initializes a new app from the exported genesis,
syncs the heights of both apps, and compares the module store data to ensure consistency.

This change also introduces necessary environment variables for the export test and fixes an issue where the znet removal was prematurely called.

* Merge branch 'master' into mehdi/export-integration-test

* Refactor for txd

* Remove wasm from ignore list

* Merge branch 'master' into mehdi/export-integration-test

* Merge branch 'master' into mehdi/export-integration-test

* Merge branch 'master' into mehdi/export-integration-test

* Merge branch 'master' into mehdi/export-integration-test

* Merge branch 'master' into mehdi/export-integration-test

* Remove recover and panic from test

* Refactors integration tests for export

Refactors the integration tests export to use the root command directly instead of relying on the external binary.
This change simplifies the test setup and makes it more robust.
Also, it fixes an issue where the test was only looking for a full node and now looks for either a full or val node to retrieve genesis.

* Merge branch 'master' into mehdi/export-integration-test

* Merge branch 'master' into mehdi/export-integration-test

* Move znet home to flags in export test

* Enable export test after upgrade integration test only

* Disable Upgrade module export test

* Move genesis export to crust

* Use lo.must in simap parse

* Merge branch 'master' into mehdi/export-integration-test

* Update crust ref

* Merge branch 'master' into mehdi/export-integration-test

* lint: remove unuded funcs

* Remove unused references

* Merge branch 'master' into mehdi/export-integration-test

* Remove unnecessary helper funcs from simap